Modify solution add/remove commands to avoid printing
Solution add_project and remove_project now take a callback which is
passed success, message, and code. This is because the function is async
and cannot return these values. The caller provides a callback which
then decides how to alert the user of success or failure.

This change is made in preparation for a legitimate UI rather than
commands, in which the print side-effect is unwanted.

This change also includes minor bug fixes, such as chunk being nil when
spawning dotnet sln add/remove, and from checking the incorrect string
for success in dotnet sln remove's output.