Change solution.slns from an array to a dictionary
slns is now a dictionary mapping a solution to its full path. This is
partly for conformity with solution.aggregate_projects and parly to
avoid duplicate solutions (for example, if one forcibly registers it
with :SolutionRegister)

This requires certain changes to command completions, as the table does
not function with vim.tbl_map, since vim.tbl_map maintains the keys of
the table, while the complete function needs to return an explicit
array. vim.tbl_map calls have been replaced with utils.tbl_map_to_arr.

Condition-gated commands counting the number of solutions now use
vim.tbl_count instead of the length operator.