Move commands to individual files
The monolithic commands.lua file was getting out of hand, so I decided
to split each entry in the commands array into its own file. This can be
considered similar to how I load plugins myself.

I would like to load them automatically in the future without having to
require each in the commands array, but if it cannot be done without
scanning the directory manually then I doubt I'll include it in the
future.