Fix incorrect usage of utils.tbl_first_matching in utils.resolve_project
utils.tbl_first_matching returns both the key and the value of the
matching element. In utils.resolve_project, the key was treated as the
value and the value was discarded, causing issues with anything using
utils.resolve_project (almost every command), as callers would receive
the path instead of the project object in certain cases.