Reimplement Nuget UI using new Textmenu class
The code backing the Nuget UI is now reimplemented as the Textmenu
class, which will allow later text-ui to reuse most of the code.

Textmenu builds upon the older Nuget UI and now supports expanding text
on enter, restricting cursor movement to each entry, and performing an
action with the selected entry.

This will later be used to implement various project and solution
editing UI.