Open Nuget Browser when calling :ProjectAddNugetDep with no arguments
:ProjectAddNugetDep now calls solution.nuget.ui.open, which allows one
to pick a dependency from the nuget browser directly.

Upon selecting a package, a vim.ui.select menu is opened which allows
one to pick which project to add the dependency. This may be improved in
the future, such as allowing one to pass a single argument to preselect
the project, or with a better ui.

Additionally this commit includes textmenu.alert, which creates a small
popup window to display an alert. This is used by the nuget browser when
selecting a package to inform the user of success or failure.