Allow :ProjectAddNugetDep to take a single project argument
Similarly to in the absence of any arguments, the nuget browser is
opened, but the project is preselected and one must not be selected upon
picking a package to add.

This required a few more changes to textmenus, mainly the addition of
"instances", where textmenus can represent an object. This is provided
to keybind callbacks as the first parameter.