Separate alert from textmenu
I thought about it, and came to the conclusion that alerts should be
their own object similar to textmenu, but shrimpler. This would later
allow for things like updating the text of an alert, such as for showing
a loading animation while waiting for a command to run.

This may require some kind of cursor locking mechanism to prevent
closing the window while running.