Rename commands for clarity
Renamed :Nuget to :NugetBrowser because 'Nuget' on its own is not very
descriptive.

Renamed :SolutionRegister to :SolutionNvimRegister to indicate
'Solution' refers to the plugin name rather than the object the command
is operating on, as is the pattern for every other command.

Renamed :SolutionLog to :SolutionNvimLog for the same reason.