Fix incorrect function injection into keymaps in textmenu.new
Keymap functions are modified to provide the requisite parameters to the
callback functions defined in keymaps passed to textmenu.new.
Previously, this was an in-place modification, which would cause keymaps
to provide incorrect parameters if a textmenu was opened with the same
keymaps used for a different textmenu.

This commit fixes this by copying the values present in the keymap into
a new table, and then modifying the function within the clone so the
original is unmodified.