Add solutionfile.textmenu and :SolutionMenu
This adds the function solutionfile.textmenu, which opens a textmenu
representing the given solutionfile and allows the addition and removal
of projects. More functionality is to come.

This includes the addition of the textmenu parameter provided to
functions defined as the callback of textmenu keymaps. Callbacks have
been updated in other locations to reflect this, and the type signature
has been updated.

:SolutionMenu simply calls solutionfile.textmenu with the provided
solution.