Rework project and solution command completion
This commit changes most project and solution completions to complete
file paths in the event no solutions or projects are available.

It may be helpful to eventually modify this to also provide file
completion if the text entered does not match any projects.

utils.complete_file has also changed to include patterns for directories
by default.