Fix multiple errors within ProjectFile
Rename the Dependency package field to pkg, to avoid issues with the
annotation conflicting with the lua package keyword.

The Dependency rel_path field used with project references was
incorrectly set, as well as not converted to the proper OS path. This is
now fixed.

ProjectFile:remove_project executed dotnet remove incorrectly.

A new helper function was introduced, remove_dependency, which takes any
dependency and determines the correct method of removal. This has been
used in :ProjectRemoveDep.