Ensure greater safety when constructing projects and solutions
This commit ensures consistency between the various project and solution
ctors, mainly with ensuring the provided path exists, has the right
extension, and is resolved to the full path.

The ctor for projects within solutions was also modified to avoid
reparsing the entire project if the project has already been registered.