Default project and solution detection on launch
A default project and solution will be detected on launch, consistently
selecting the same project and solution. This will later come into use
with commands which take only a single project or solution argument.

The default project and solution for a given root directory can be set,
and will be persisted to the user's data directory. These can be set
using two new commands, :ProjectSetDefault and :SolutionSetDefault.

I've attempted to make saving and loading as resilient as possible, but
there are likely still ways to break it :).