Change signature of on_exit in outputs to match utils.spawn_proc
Since the functions in each output in solution.output spawn processes
similar to utils.spawn_proc, the on_exit was changed to match.

This entailed changing the behavior of the term output, which now closes
after every command run through it so the on_exit callback can be
properly used.