Fix incorrectly determined full path on some projects
Projects found in a solution file not in the cwd (so in some folder
further down in the project) would have incorrectly determined full
paths due to an assumption that project paths were relative to the cwd.

Also avoids crashing when unable to read text from a project that
doesn't exist.